@@ -1,7 +1,10 @@
-require("./bootstrap");
+import Vue from 'vue';
+import VeeValidate from 'vee-validate';
 
-window.Vue = require("vue");
-window.VeeValidate = require("vee-validate");
+import './bootstrap';
+
+window.Vue = Vue;
+window.VeeValidate = VeeValidate;
 
 Vue.use(VeeValidate);
 Vue.prototype.$http = axios
