@@ -1,15 +1,23 @@
-window.jQuery = window.$ = $ = require("jquery");
-window.Vue = require("vue");
-window.VeeValidate = require("vee-validate");
-window.axios = require("axios");
+import $ from 'jquery';
+import Vue from 'vue';
+import VeeValidate from 'vee-validate';
+import axios from 'axios';
+import VueSlider from 'vue-slider-component';
+import accounting from 'accounting';
+
+import ImageSlider from './components/image-slider';
+import { messages as localeMessages } from './lang/locales';
+
+window.jQuery = window.$ = $;
+window.Vue = Vue;
+window.VeeValidate = VeeValidate;
+window.axios = axios;
 require("./bootstrap");
 require("ez-plus/src/jquery.ez-plus.js");
-var accounting = require('accounting');
-locales = require("./lang/locales.js");
 
 Vue.use(VeeValidate, {
     dictionary: {
-        ar: { messages: locales.messages.ar }
+        ar: { messages: localeMessages.ar }
     }
 });
 
@@ -17,8 +25,8 @@ Vue.prototype.$http = axios
 
 window.eventBus = new Vue();
 
-Vue.component("image-slider", require("./components/image-slider.vue"));
-Vue.component("vue-slider", require("vue-slider-component"));
+Vue.component("image-slider", ImageSlider);
+Vue.component("vue-slider", VueSlider);
 Vue.filter('currency', function (value, argument) {
     return accounting.formatMoney(value, argument);
 })
